refactor : [STEP3] 같은 주기로 생성되고 제거되는 menu, menugroup 합치기
