+++ src/main/java/kitchenpos/table/domain/TableGroupRepository.java
@@ -1,0 +1,5 @@
+package kitchenpos.table.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {}
