--- src/main/java/kitchenpos/tablegroup/dto/TableGroupRequest.java
@@ -1,32 +1,0 @@
-package kitchenpos.tablegroup.dto;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-import kitchenpos.table.dto.OrderTableRequest;
-
-public class TableGroupRequest {
-
-    private List<OrderTableRequest> orderTables;
-
-    protected TableGroupRequest() {}
-
-    public TableGroupRequest(List<OrderTableRequest> orderTables) {
-        this.orderTables = orderTables;
-    }
-
-    public List<Long> getOrderTableIds() {
-        return orderTables.stream()
-            .map(orderTableRequest -> orderTableRequest.getId())
-            .collect(Collectors.toList());
-    }
-
-    public List<OrderTableRequest> getOrderTables() {
-        return orderTables;
-    }
-
-    public void setOrderTables(List<OrderTableRequest> orderTables) {
-        this.orderTables = orderTables;
-    }
-
-}
